total strategy rewrite
I found that even knowing the knock, I was only able to get it right
about once every 5-10 tries. And I think a large part of the problem was
the stupid locality-sensitive hash, that just wasn't very good.

So this take a totally different approach: we detect a sequence of
knocks followed by silence, and then compare the sequence to our desired
one. It seems to work much better in practice.

And a nice side effect is that the audio processing is totally split
from the sequence-matching logic, making both stages much easier to
reason about.